Create supercomputer.cpp
The Goal
In the Computer2000 data center, you are responsible for planning the usage of a supercomputer for scientists. ​Therefore you've decided to organize things a bit by planning everybody’s tasks. The logic is simple: the higher the number of calculations which can be performed, the more people you can satisfy.
 
Rules
Scientists give you the starting day of their calculation and the number of consecutive days they need to reserve the calculator.

Game Input

Input
Line 1: The number N of calculations
The N following lines: on each line, the starting day J and the duration D of reservation, separated by a blank space.

Output
The maximum number of calculations that can be carried out.

Constraints
0 < N < 100000
0 < J < 1000000
0 < D < 1000